+++ internal/app/forklift/cli/pallets-deployments.go
@@ -1,15 +1,20 @@
 package cli
 
 import (
+	"bytes"
 	"fmt"
+	"os"
 	"path"
+	"path/filepath"
 	"sort"
 
 	dct "github.com/compose-spec/compose-go/v2/types"
 	"github.com/pkg/errors"
+	"gopkg.in/yaml.v3"
 
 	"github.com/PlanktoScope/forklift/internal/app/forklift"
 	"github.com/PlanktoScope/forklift/pkg/core"
+	"github.com/PlanktoScope/forklift/pkg/structures"
 )
 
 // Print
@@ -31,7 +36,7 @@ func PrintDeplInfo(
 	depl, err := pallet.LoadDepl(deplName)
 	if err != nil {
 		return errors.Wrapf(
-			err, "couldn't find package deployment specification %s in pallet %s",
+			err, "couldn't find package deployment declaration %s in pallet %s",
 			deplName, pallet.FS.Path(),
 		)
 	}
@@ -247,7 +252,7 @@ func PrintDeplPkgPath(
 	depl, err := pallet.LoadDepl(deplName)
 	if err != nil {
 		return errors.Wrapf(
-			err, "couldn't find package deployment specification %s in pallet %s",
+			err, "couldn't find package deployment declaration %s in pallet %s",
 			deplName, pallet.FS.Path(),
 		)
 	}
@@ -259,5 +264,264 @@ func PrintDeplPkgPath(
 		return errors.Errorf("package deployment %s is not enabled!", depl.Name)
 	}
 	fmt.Println(resolved.Pkg.FS.Path())
+	return nil
+}
+
+// Add
+
+func AddDepl(
+	indent int, pallet *forklift.FSPallet, pkgLoader forklift.FSPkgLoader,
+	deplName, pkgPath string, features []string, disabled, force bool,
+) error {
+	disabledString := ""
+	if disabled {
+		disabledString = "disabled "
+	}
+	featuresString := ""
+	if len(features) > 0 {
+		featuresString = fmt.Sprintf(" (with feature flags: %+v)", features)
+	}
+	IndentedPrintf(
+		indent, "Adding %spackage deployment %s for %s%s...\n",
+		disabledString, deplName, pkgPath, featuresString,
+	)
+	depl := forklift.Depl{
+		Name: deplName,
+		Def: forklift.DeplDef{
+			Package:  pkgPath,
+			Features: features,
+			Disabled: disabled,
+		},
+	}
+
+	if err := checkDepl(pallet, pkgLoader, depl); err != nil {
+		if !force {
+			return errors.Wrap(
+				err, "package deployment has invalid settings; to skip this check, enable the --force flag",
+			)
+		}
+		IndentedPrintf(indent, "Warning: package deployment has invalid settings: %s", err.Error())
+	}
+
+	if err := writeDepl(pallet, depl); err != nil {
+		return errors.Wrapf(err, "couldn't save deployment %s", deplName)
+	}
+	return nil
+}
+
+func checkDepl(
+	pallet *forklift.FSPallet, pkgLoader forklift.FSPkgLoader, depl forklift.Depl,
+) error {
+	pkg, _, err := forklift.LoadRequiredFSPkg(pallet, pkgLoader, depl.Def.Package)
+	if err != nil {
+		return errors.Wrapf(
+			err, "couldn't resolve package path %s to a package using the pallet's repo requirements",
+			depl.Def.Package,
+		)
+	}
+
+	allowedFeatures := pkg.Def.Features
+	unrecognizedFeatures := make([]string, 0, len(depl.Def.Features))
+	for _, name := range depl.Def.Features {
+		if _, ok := allowedFeatures[name]; !ok {
+			unrecognizedFeatures = append(unrecognizedFeatures, name)
+		}
+	}
+	if len(unrecognizedFeatures) > 0 {
+		return errors.Errorf("unrecognized feature flags: %+v", unrecognizedFeatures)
+	}
+	return nil
+}
+
+func writeDepl(pallet *forklift.FSPallet, depl forklift.Depl) error {
+	deplsFS, err := pallet.GetDeplsFS()
+	if err != nil {
+		return err
+	}
+	deplPath := path.Join(deplsFS.Path(), fmt.Sprintf("%s.deploy.yml", depl.Name))
+	buf := bytes.Buffer{}
+	encoder := yaml.NewEncoder(&buf)
+	const yamlIndent = 2
+	encoder.SetIndent(yamlIndent)
+	if err = encoder.Encode(depl.Def); err != nil {
+		return errors.Wrapf(err, "couldn't marshal package deployment for %s", deplPath)
+	}
+	if err := forklift.EnsureExists(filepath.FromSlash(path.Dir(deplPath))); err != nil {
+		return errors.Wrapf(
+			err, "couldn't make directory %s", filepath.FromSlash(path.Dir(deplPath)),
+		)
+	}
+	const perm = 0o644 // owner rw, group r, public r
+	if err := os.WriteFile(filepath.FromSlash(deplPath), buf.Bytes(), perm); err != nil {
+		return errors.Wrapf(
+			err, "couldn't save deployment declaration to %s", filepath.FromSlash(deplPath),
+		)
+	}
+	return nil
+}
+
+// Remove
+
+func RemoveDepls(indent int, pallet *forklift.FSPallet, deplNames []string) error {
+	fmt.Printf("Removing package deployments from %s...\n", pallet.FS.Path())
+	for _, deplName := range deplNames {
+		deplsFS, err := pallet.GetDeplsFS()
+		if err != nil {
+			return err
+		}
+		deplPath := path.Join(deplsFS.Path(), fmt.Sprintf("%s.deploy.yml", deplName))
+		if err = os.RemoveAll(deplPath); err != nil {
+			return errors.Wrapf(
+				err, "couldn't remove package deployment %s, at %s", deplName, deplPath,
+			)
+		}
+	}
+	// TODO: maybe it'd be better to remove everything we can remove and then report errors at the
+	// end?
+	return nil
+}
+
+// Set Package
+
+func SetDeplPkg(
+	indent int, pallet *forklift.FSPallet, pkgLoader forklift.FSPkgLoader,
+	deplName, pkgPath string, force bool,
+) error {
+	IndentedPrintf(
+		indent, "Setting package deployment %s to deploy package %s...\n", deplName, pkgPath,
+	)
+	depl, err := pallet.LoadDepl(deplName)
+	if err != nil {
+		return errors.Wrapf(
+			err, "couldn't find package deployment declaration %s in pallet %s",
+			deplName, pallet.FS.Path(),
+		)
+	}
+
+	depl.Def.Package = pkgPath
+	// We want to check both the package path and the feature flags for validity, since changing the
+	// package path could change the allowed feature flags:
+	if err := checkDepl(pallet, pkgLoader, depl); err != nil {
+		if !force {
+			return errors.Wrap(
+				err, "package deployment has invalid settings; to skip this check, enable the --force flag",
+			)
+		}
+		IndentedPrintf(indent, "Warning: package deployment has invalid settings: %s", err.Error())
+	}
+
+	if err := writeDepl(pallet, depl); err != nil {
+		return errors.Wrapf(err, "couldn't save updated deployment declaration %s", depl.Name)
+	}
+	return nil
+}
+
+// Add Feature
+
+func AddDeplFeat(
+	indent int, pallet *forklift.FSPallet, pkgLoader forklift.FSPkgLoader,
+	deplName string, features []string, force bool,
+) error {
+	IndentedPrintf(indent, "Enabling features %+v in package deployment %s...\n", features, deplName)
+	depl, err := pallet.LoadDepl(deplName)
+	if err != nil {
+		return errors.Wrapf(
+			err, "couldn't find package deployment declaration %s in pallet %s",
+			deplName, pallet.FS.Path(),
+		)
+	}
+	resolved, err := forklift.ResolveDepl(pallet, pkgLoader, depl)
+	if err != nil {
+		return errors.Wrapf(err, "couldn't resolve package deployment %s", depl.Name)
+	}
+
+	existingFeatures := make(structures.Set[string])
+	for _, name := range depl.Def.Features {
+		existingFeatures.Add(name)
+	}
+	allowedFeatures := resolved.Pkg.Def.Features
+	unrecognizedFeatures := make([]string, 0, len(features))
+	newFeatures := make([]string, 0, len(features))
+	for _, name := range features {
+		if _, ok := allowedFeatures[name]; !ok {
+			unrecognizedFeatures = append(unrecognizedFeatures, name)
+		}
+		if existingFeatures.Has(name) {
+			continue
+		}
+		newFeatures = append(newFeatures, name)
+		existingFeatures.Add(name) // suppress duplicates in the input features list
+	}
+	if len(unrecognizedFeatures) > 0 {
+		err := errors.Errorf(
+			"feature flags %+v are allowed by package %s; to skip this check, enable the --force flag",
+			unrecognizedFeatures, depl.Def.Package,
+		)
+		if !force {
+			return err
+		}
+		IndentedPrintf(indent, "Warning: %s", err.Error())
+	}
+
+	depl.Def.Features = append(depl.Def.Features, newFeatures...)
+	if err := writeDepl(pallet, depl); err != nil {
+		return errors.Wrapf(err, "couldn't save updated deployment declaration %s", depl.Name)
+	}
+	return nil
+}
+
+// Remove Feature
+
+func RemoveDeplFeat(
+	indent int, pallet *forklift.FSPallet, deplName string, features []string,
+) error {
+	IndentedPrintf(indent, "Disabling features %+v in package deployment %s...\n", features, deplName)
+	depl, err := pallet.LoadDepl(deplName)
+	if err != nil {
+		return errors.Wrapf(
+			err, "couldn't find package deployment declaration %s in pallet %s",
+			deplName, pallet.FS.Path(),
+		)
+	}
+
+	removedFeatures := make(structures.Set[string])
+	for _, name := range features {
+		removedFeatures.Add(name)
+	}
+	newFeatures := make([]string, 0, len(depl.Def.Features))
+	for _, name := range depl.Def.Features {
+		if removedFeatures.Has(name) {
+			continue
+		}
+		newFeatures = append(newFeatures, name)
+	}
+
+	depl.Def.Features = newFeatures
+	if err := writeDepl(pallet, depl); err != nil {
+		return errors.Wrapf(err, "couldn't save updated deployment declaration %s", depl.Name)
+	}
+	return nil
+}
+
+// Set Disabled
+
+func SetDeplDisabled(indent int, pallet *forklift.FSPallet, deplName string, disabled bool) error {
+	if disabled {
+		IndentedPrintf(indent, "Disabling package deployment %s...\n", deplName)
+	} else {
+		IndentedPrintf(indent, "Enabling package deployment %s...\n", deplName)
+	}
+	depl, err := pallet.LoadDepl(deplName)
+	if err != nil {
+		return errors.Wrapf(
+			err, "couldn't find package deployment declaration %s in pallet %s",
+			deplName, pallet.FS.Path(),
+		)
+	}
+
+	depl.Def.Disabled = disabled
+	if err := writeDepl(pallet, depl); err != nil {
+		return errors.Wrapf(err, "couldn't save updated deployment declaration %s", depl.Name)
+	}
 	return nil
 }